Move back history code
