+++ cmd/genji/shell/shell.go
@@ -1,6 +1,7 @@
 package shell
 
 import (
+	"bufio"
 	"context"
 	"encoding/json"
 	"errors"
@@ -8,6 +9,7 @@ import (
 	"io/ioutil"
 	"os"
 	"os/signal"
+	"path/filepath"
 	"strings"
 	"sync"
 	"syscall"
@@ -352,6 +354,65 @@ func (sh *Shell) loadCommandSuggestions() {
 		}
 	}
 	sh.cmdSuggestions = suggestions
+}
+
+func (sh *Shell) loadHistory() ([]string, error) {
+	if _, ok := os.LookupEnv("NO_HISTORY"); ok {
+		return nil, nil
+	}
+	homeDir, err := os.UserHomeDir()
+	if err != nil {
+		return nil, err
+	}
+
+	fname := filepath.Join(homeDir, historyFilename)
+
+	_, err = os.Stat(fname)
+	if err != nil {
+		return nil, nil
+	}
+
+	f, err := os.Open(fname)
+	if err != nil {
+		return nil, nil
+	}
+	defer f.Close()
+
+	var history []string
+	s := bufio.NewScanner(f)
+	for s.Scan() {
+		history = append(history, s.Text())
+	}
+
+	return history, s.Err()
+}
+
+func (sh *Shell) dumpHistory() error {
+	if _, ok := os.LookupEnv("NO_HISTORY"); ok {
+		return nil
+	}
+	homeDir, err := os.UserHomeDir()
+	if err != nil {
+		return err
+	}
+
+	fname := filepath.Join(homeDir, historyFilename)
+
+	f, err := os.OpenFile(fname, os.O_WRONLY|os.O_CREATE|os.O_APPEND, 0600)
+	if err != nil {
+		return err
+	}
+	defer f.Close()
+
+	w := bufio.NewWriter(f)
+	for _, h := range sh.history {
+		_, err = w.WriteString(h + "\n")
+		if err != nil {
+			return err
+		}
+	}
+
+	return w.Flush()
 }
 
 // executeInput stores user input in the history and executes it.
